+++ packages/lui/src/lib/components/content/render/LuiContentRenderNumber.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+  interface Props {
+    value: number;
+  }
+
+  let { value }: Props = $props();
+</script>
+
+<p>{value}</p>