+++ packages/lui/src/lib/components/content/render/LuiContentRenderObject.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+  interface Props {
+    value: object;
+  }
+
+  let { value }: Props = $props();
+</script>
+
+<span>{JSON.stringify(value, null, 22)}</span>