+++ packages/lui/src/lib/components/content/render/LuiContentRenderString.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+  interface Props {
+    value: string;
+  }
+
+  let { value }: Props = $props();
+</script>
+
+<p>{value}</p>