+++ packages/lui/src/lib/index.ts
@@ -4,3 +4,5 @@ export { default as LuiDebugger } from './pages/LuiDebugger.svelte';
 export { default as LuiContentCard } from './components/content/LuiContentCard.svelte';
 export { default as LuiContentHint } from './components/content/LuiContentHint.svelte';
 export { default as LuiContext } from './LuiContext.svelte';
+
+export type { LuiContentRenderRegistry } from './components/content/render/LuiContentRenderRegistry';