+++ packages/lui/src/lib/util/LuiError.ts
@@ -3,3 +3,5 @@ export abstract class LuiError extends Error {}
 export class EngineNotFoundError extends LuiError {}
 
 export class IntrospectionNotFoundError extends LuiError {}
+
+export class ContentRenderRegistryNotFoundError extends LuiError {}