+++ packages/lui/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @123ishatest/lui
 
+## 1.0.3
+
+### Patch Changes
+
+- Fix LuiContentHint not being in scope properly
+
+## 1.0.2
+
+### Patch Changes
+
+- Add ContentRegistry and export Node types
+- Updated dependencies
+  - @123ishatest/ludiek@0.9.1
+
 ## 1.0.1
 
 ### Patch Changes